fastapi + docker/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
Pillow
python-multipart
